# chore(mwpw-1234): boost Lighthouse performance to 99 by swapping to Preact & enabling gzip/Brotli server