--- a/index.html
+++ b/index.html
@@ -47,7 +47,7 @@
 
 
 
-    <script type="text/javascript" src="./dist/main.source.js"></script>
+    <script type="text/javascript" src="./dist/main.min.js"></script>
 
     <script type="text/javascript">
         var config = {