--- a/index.html
+++ b/index.html
@@ -10,6 +10,11 @@
     <meta name="template" content="" />
     <meta name="description" content="Card Artboard 2" />
 
+    <!-- Preconnect to Adobe domains and preload CSS to improve LCP -->
+    <link rel="preconnect" href="https://www.adobe.com" crossorigin>
+    <link rel="dns-prefetch" href="https://www.adobe.com">
+    <link rel="preload" href="./dist/app.css" as="style">
+
     <link rel="stylesheet" href="./dist/app.css" type="text/css">
     <link rel="stylesheet" href="./dependencies/typekit.css" type="text/css">
     <link rel="stylesheet" href="./dependencies/dexter/publish.css" type="text/css">
@@ -45,7 +50,12 @@
         </div>
     </div>
 
-
+    <link
+            rel="preload"
+            as="image"
+            href="https://www.adobe.com/content/dam/www/us/en/max/2021/speakers/S907-Deep-Blizzard_1628658923394001pGLY.jpg"
+            type="image/jpeg"
+    />
 
     <script type="text/javascript" src="./dist/main.min.js"></script>