--- a/jest.config.js
+++ b/jest.config.js
@@ -2,6 +2,8 @@
 // https://jestjs.io/docs/en/configuration.html
 
 module.exports = {
+    // Module name mappings (none by default)
+    moduleNameMapper: {},
     // Automatically clear mock calls and instances between every test
     clearMocks: true,
     collectCoverage: true,