--- a/package.json
+++ b/package.json
@@ -44,6 +44,7 @@
     "jest-cli": "^23.6.0",
     "prop-types": "^15.6.2",
     "react": "^16.12.0",
+    "preact": "^10.0.0",
     "react-dom": "^16.12.0",
     "react-dom-components": "^3.0.0",
     "react-if": "2.2.2",