--- a/sourcemaps.config.js
+++ b/sourcemaps.config.js
@@ -32,6 +32,14 @@ module.exports = {
         path: path.resolve(__dirname, 'dist'),
     },
     resolve: {
+        // Alias React imports to Preact for compatibility in development sourcemap build
+        alias: {
+            react: 'preact/compat',
+            'react-dom': 'preact/compat',
+            // alias test-utils and JSX runtime if used
+            'react-dom/test-utils': 'preact/test-utils',
+            'react/jsx-runtime': 'preact/jsx-runtime'
+        },
         extensions: ['.js', '.jsx'],
     },
     module: {