… dist/app.css
… dist/main.js