# What's Changed

- Swapped out React for Preact to reduce bundle size (~172 KB of JS eliminated via tree-shaking & smaller runtime)
- Adopted the existing `serve` CLI (`npm run serve`) as our local static server to:
  - Enable on-the-fly gzip and Brotli compression (saved ~758 KB of text)
  - Emit optimal `Cache-Control` headers for immutable assets
  - Support SPA fallback (`--single`) for index.html routing
- Documented the new recommended local-dev workflow in README

## Why This Matters

- **Performance**: Preact's smaller footprint + compressed assets moved our mobile Lighthouse Performance score from **62** → **99**
- **Developer DX**: Zero-config static server ("serve") gives us compression, caching, and SPA support without custom Python scripts or WebStorm hacks
- **Maintainability**: We keep using our existing Webpack build and npm scripts, so there's no new build tool or CI complexity

## Before & After

| Metric                     | Before          | After          | Savings       |
|----------------------------|-----------------|----------------|---------------|
| Lighthouse Performance     | 62 / 100        | 99 / 100       | +37 pts       |
| Bundle size (main.js)      | ~200 KB gzipped | ~28 KB gzipped | –172 KB       |
| Text compression potential | ~758 KB wasted  | ~0 KB wasted   | ~758 KB saved |

## How to Verify

1. `npm install`
2. `npm run serve -- --single --brotli --listen 8080`
3. Point your browser (or Lighthouse CLI) at `http://localhost:8080`
4. Confirm Performance ≥ 99, FCP/LCP < 1 s, 0 TBT, 0 CLS

## Docs

- Added "Development" section to **README.md** explaining the `serve` flags for compression, HTTP/2 readiness, and SPA fallback.

After preact migration:
<img width="943" alt="image" src="https://github.com/user-attachments/assets/63d4ab0a-db24-4cda-b1ad-130a3437e4a9" />